refactor: 패키지 이름 domain으로 변경
