+++ src/main/java/tosiltosil/backend/common/logging/domain/InfoLog.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.logging.dto;
+package tosiltosil.backend.common.logging.domain;
 
 import lombok.extern.slf4j.Slf4j;
